# api/openapi.yaml
openapi: 3.1.0
info:
  title: QianuHub 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/analytics/orders":
    get:
      operationId: getApiV1AnalyticsOrders
      summary: getApiV1AnalyticsOrders
      security: [{cookieAuth: []}]
      parameters:
        - name: start_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: end_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: timezone_offset_minutes
          in: query
          required: false
          schema: {type: integer}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/orders":
    get:
      operationId: getApiV1Orders
      summary: getApiV1Orders
      security: [{cookieAuth: []}]
      parameters:
        - name: cookie_id
          in: query
          required: false
          schema: {type: string}
        - name: status
          in: query
          required: false
          schema: {type: string}
        - name: search
          in: query
          required: false
          schema: {type: string}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 500}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/orders/{order_id}":
    delete:
      operationId: deleteApiV1OrdersByorder_id
      summary: deleteApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getApiV1OrdersByorder_id
      summary: getApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putApiV1OrdersByorder_id
      summary: putApiV1OrdersByorder_id
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/orders/{order_id}/refresh":
    post:
      operationId: postApiV1OrdersByorder_idRefresh
      summary: postApiV1OrdersByorder_idRefresh
      parameters:
          - name: order_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSingleRefreshResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/orders/import":
    post:
      operationId: postApiV1OrdersImport
      summary: 人工插入订单已永久移除
      deprecated: true
      x-contract-success-exception:
        kind: permanently_disabled
      security: [{cookieAuth: []}]
      responses:
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: 人工插入订单已移除，请使用订单同步
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/orders/manual-ship":
    post:
      operationId: postApiV1OrdersManualShip
      summary: postApiV1OrdersManualShip
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualShipRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBatchResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      description: 统一响应的稳定字段；业务成功字段由各阶段 operation schema 收紧。
...
    OrderDTO:
      type: object
      required: [order_id, item_id, item_title, item_image, buyer_id, quantity, amount, order_status, status, cookie_id, created_at]
      properties:
        order_id: {type: string}
        item_id: {type: string}
        item_title: {type: string}
        item_image: {type: string}
        buyer_id: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        quantity: {type: string}
        amount: {type: string}
        order_status: {type: string}
        status: {type: string}
        cookie_id: {type: string}
        is_bargain: {type: integer}
        system_shipped: {type: boolean}
        receiver_name: {type: string}
        receiver_phone: {type: string}
        receiver_address: {type: string}
        receiver_city: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    OrderListResponse:
      type: object
      required: [success, data, total, page, page_size, total_pages]
      properties:
        success: {type: boolean}
        data: {type: array, items: {$ref: '#/components/schemas/OrderDTO'}}
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        total_pages: {type: integer}
    OrderDetailResponse:
      allOf:
        - {$ref: '#/components/schemas/OrderDTO'}
        - type: object
          required: [success, data]
          properties:
            success: {type: boolean}
            data: {$ref: '#/components/schemas/OrderDTO'}
    OrderUpdateRequest:
      type: object
      properties:
        order_status: {type: string}
        status: {type: string}
        item_id: {type: string}
        buyer_id: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        quantity: {oneOf: [{type: string}, {type: number}]}
        amount: {oneOf: [{type: string}, {type: number}]}
        receiver_name: {type: string}
        receiver_phone: {type: string}
        receiver_address: {type: string}
        receiver_city: {type: string}
        chat_id: {type: string}
        system_shipped: {type: boolean}
        item_title: {type: string}
...
    OrderRefreshDetail:
      type: object
      required: [quantity, spec_name, spec_value, order_status, amount]
      properties:
        quantity: {type: string}
        spec_name: {type: string}
        spec_value: {type: string}
        order_status: {type: string}
        amount: {type: string}
    OrderSingleRefreshResponse:
      type: object
      required: [success, message, order]
      properties:
        success: {type: boolean}
        message: {type: string}
        order: {$ref: '#/components/schemas/OrderRefreshDetail'}
...
    ManualShipRequest:
      type: object
      required: [order_ids, ship_mode]
      properties:
        order_ids: {type: array, items: {type: string}}
        ship_mode: {type: string, enum: [status_only, full_delivery]}
    OrderBatchResult:
      type: object
      required: [message]
      properties:
        order_id: {type: string}
        status: {type: string}
        success: {type: boolean}
        message: {type: string}
        cookie_id: {type: string}
        reconciliation_id: {type: string}
        reconciliation_warning: {type: string}
        stage: {type: string}
    OrderBatchResponse:
      type: object
      required: [partial_failure, message, success_count, failed_count, results]
      properties:
        partial_failure: {type: boolean}
        message: {type: string}
        total: {type: integer}
        success_count: {type: integer}
        failed_count: {type: integer}
        results: {type: array, items: {$ref: '#/components/schemas/OrderBatchResult'}}
...
    AnalyticsResponse:
      type: object
      required: [revenue_stats, daily_stats, status_stats, city_stats, item_stats]
      properties:
        revenue_stats: {type: object, required: [total_orders, total_amount, avg_amount, unique_buyers, unique_items], properties: {total_orders: {type: integer}, total_amount: {type: number}, avg_amount: {type: number}, unique_buyers: {type: integer}, unique_items: {type: integer}}}
        daily_stats: {type: array, items: {type: object, required: [date, order_count, amount], properties: {date: {type: string}, order_count: {type: integer}, amount: {type: number}}}}
        status_stats: {type: array, items: {type: object, required: [status, count, amount], properties: {status: {type: string}, count: {type: integer}, amount: {type: number}}}}
        city_stats: {type: array, items: {type: object, required: [city, order_count, total_amount], properties: {city: {type: string}, order_count: {type: integer}, total_amount: {type: number}}}}
        item_stats: {type: array, items: {type: object, required: [item_id, order_count, total_amount, avg_amount], properties: {item_id: {type: string}, order_count: {type: integer}, total_amount: {type: number}, avg_amount: {type: number}}}}
...
    OperationResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        requires_relogin:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true